[BUGFIX] Assignments are not available in cached context
The assignments have been defined in ext_tables.php which is not
parsed in a cached frontend rendering context (below TYPO3 6.2).
Thus, the registration process needs to happen earlier during
ext_localconf.php processing. Besides that, the configuration
service in local GeneralUtility has to be initialized without
relying on the Extbase ObjectManager.